Restore past Password Manager settings behavior from M125 or older
As of M126, Play services (GMSCore) is used as backend by default for
the saved passwords in Password Manager settings, rendering the previous
backend used to save password inaccessible. This commit restores access
to the password saved in older versions.

In Chrome, it is labeled as Google Password Manager,
which uses GMSCore to store the saved passwords.
As of https://chromium-review.googlesource.com/c/chromium/src/+/5493819m
this backend is planned to be removed in November 2024.